Simplify calling of events and properties functions and move back under lifecycle.
