+++ src/lifecycle/attribute.js
@@ -3,7 +3,6 @@ import data from '../util/data';
 export default function (opts) {
   var callback = opts.attribute || function () {};
 
-  /* jshint expr: true */
   return function (name, oldValue, newValue) {
     var info = data(this);
     var attributeToPropertyMap = info.attributeToPropertyMap || {};
