+++ test/unit.js
@@ -2,11 +2,9 @@ import './boot';
 import './unit/api/compose';
 import './unit/api/create';
 import './unit/api/emit';
-import './unit/api/events';
 import './unit/api/fragment';
 import './unit/api/global';
 import './unit/api/init';
-import './unit/api/properties';
 import './unit/attributes';
 import './unit/constructor';
 import './unit/document-observer';
@@ -16,6 +14,8 @@ import './unit/ignoring';
 import './unit/lifecycle';
 import './unit/lifecycle-scenarios';
 import './unit/lifecycle/created';
+import './unit/lifecycle/events';
+import './unit/lifecycle/properties';
 import './unit/lifecycle/ready';
 import './unit/registration';
 import './unit/registry';
