+++ test/unit/lifecycle/events.js
