+++ test/unit/lifecycle/properties.js
